+++ src/main/kotlin/com/github/rishabjaiswal/dartassetmanager/toolWindow/PackageInfo.kt
@@ -5,7 +5,8 @@ import com.intellij.openapi.vfs.VirtualFile
 data class PackageInfo(
     val name: String,
     val directory: VirtualFile?,
-    val pubspecFile: VirtualFile?
+    val pubspecFile: VirtualFile?,
+    val bundledAssets: PubspecAssets,
 ) {
     override fun toString(): String = name
 }